chore: isolate more region and reusable section from home component
