+++ sites/public/src/components/home/HomeRegions.module.scss
@@ -1,0 +1,7 @@
+.region-card {
+  .region-card-name {
+    padding: var(--seeds-s6);
+    color: var(--seeds-color-primary);
+    text-decoration: underline;
+  }
+}
