+++ sites/public/src/components/home/HomeSection.module.scss
@@ -1,0 +1,18 @@
+.section-header {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  gap: var(--seeds-s2_5);
+  margin-top: var(--seeds-spacer-container);
+  margin-bottom: var(--seeds-spacer-content);
+
+  .section-header-icon {
+    color: var(--seeds-color-black)
+  }
+
+  .section-header-title {
+    color: var(--seeds-text-color-darker);
+    font-family: var(--seeds-font-sans);
+    font-weight: var(--seeds-font-weight-semibold);
+  }
+}
